requirements.txt
streamlit==1.30.0
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.3
lancedb==0.14.0
ipykernel==6.29.5
nltk==3.9.1
langchain==0.3.17
langchain-community==0.3.16
sentence_transformers==3.4.1